refactor(lab-6): add versioning to the project.

+++ blog-common/src/main/java/com/kratosgado/blog/dtos/response/CategoryResponse.java
@@ -5,5 +5,5 @@ public record CategoryResponse(
     String name,
     String slug,
     String description,
-    Integer postCount) {
+    Long postCount) {
 }
